[FEAT] remove old actions
