+++ src/features/world/ui/TradeCompleted.tsx
@@ -14,7 +14,6 @@ import token from "assets/icons/token_2.png";
 import { Button } from "components/ui/Button";
 import { OuterPanel } from "components/ui/Panel";
 import { useAppTranslation } from "lib/i18n/useAppTranslations";
-import { hasFeatureAccess } from "lib/flags";
 
 interface Props {
   farmId: number;
@@ -72,10 +71,6 @@ export const TradeCompleted: React.FC<Props> = ({ mmoService, farmId }) => {
                   onClick={() => {
                     gameService.send("trade.received", {
                       tradeId: trade?.tradeId,
-                      beta: hasFeatureAccess(
-                        gameState.context.state,
-                        "TRADING_REVAMP"
-                      ),
                     });
                     gameService.send("SAVE");
 
@@ -87,11 +82,7 @@ export const TradeCompleted: React.FC<Props> = ({ mmoService, farmId }) => {
 
                 <div className="flex items-center mt-3 mr-0.5">
                   <img src={token} className="h-6 mr-1" />
-                  <p className="text-xs">
-                    {hasFeatureAccess(gameState.context.state, "TRADING_REVAMP")
-                      ? (sold.sfl * 0.9).toFixed(2)
-                      : sold.sfl}
-                  </p>
+                  <p className="text-xs">{(sold.sfl * 0.9).toFixed(2)}</p>
                 </div>
               </div>
             </div>
