+++ src/features/world/ui/npcs/TradingBoard.tsx
@@ -6,7 +6,6 @@ import { BuyPanel } from "../trader/BuyPanel";
 import { Trade } from "features/bumpkins/components/Trade";
 import { Context } from "features/game/GameProvider";
 import { useActor } from "@xstate/react";
-import { hasFeatureAccess } from "lib/flags";
 
 import tradeIcon from "assets/icons/trade.png";
 
@@ -25,24 +24,7 @@ export const TradingBoard: React.FC<Props> = ({ onClose }) => {
   } = gameState;
 
   const notCloseable = gameService.state.matches("fulfillTradeListing");
-  const BetaText =
-    "Beta testers are working hard to make this feature available to you soon!";
 
-  if (!hasFeatureAccess(state, "TRADING_REVAMP")) {
-    return (
-      <CloseButtonPanel
-        onClose={notCloseable ? undefined : onClose}
-        tabs={[
-          { icon: SUNNYSIDE.icons.search, name: t("buy") },
-          { icon: tradeIcon, name: t("sell") },
-        ]}
-        setCurrentTab={setTab}
-        currentTab={tab}
-      >
-        <div className="p-1 text-sm">{BetaText}</div>
-      </CloseButtonPanel>
-    );
-  }
   return (
     <CloseButtonPanel
       onClose={notCloseable ? undefined : onClose}
