+++ src/lib/flags.ts
@@ -17,7 +17,6 @@ type FeatureName =
   | "PORTALS"
   | "DEQUIPPER"
   | "CHESTS"
-  | "TRADING_REVAMP"
   | "PWA_INSTALL";
 
 // Used for testing production features
@@ -33,7 +32,6 @@ const featureFlags: Record<FeatureName, FeatureFlag> = {
   JEST_TEST: defaultFeatureFlag,
   DEQUIPPER: defaultFeatureFlag,
   CHESTS: defaultFeatureFlag,
-  TRADING_REVAMP: testnetFeatureFlag,
   PWA_INSTALL: testnetFeatureFlag,
 };
